+++ src/test/converter2/issues/gh2954.ts
@@ -1,0 +1,11 @@
+export type AliasA = Readonly<Record<string, string>>;
+
+export type AliasB<T> = Readonly<Record<string, T>>;
+
+export type AliasC = Readonly<{}>;
+
+export interface InterfaceA {
+    propertyA: AliasA;
+    propertyB: AliasB<string>;
+    propertyC: AliasC;
+}